+++ test/unit/tracing/tracer.spec.js
@@ -164,7 +164,10 @@ describe("Test Tracer", () => {
 
 			const tracer = new Tracer(broker, {
 				enabled: true,
-				exporter: "Exporter1"
+				exporter: "Exporter1",
+				defaultTags: {
+					a: 5
+				}
 			});
 
 			tracer.init();
@@ -174,6 +177,8 @@ describe("Test Tracer", () => {
 			expect(Exporters.resolve).toHaveBeenNthCalledWith(1, "Exporter1");
 			expect(fakeExporter.init).toBeCalledTimes(1);
 			expect(fakeExporter.init).toHaveBeenNthCalledWith(1, tracer);
+
+			expect(tracer.defaultTags).toEqual({ a: 5 });
 		});
 
 		it("should initialize exporters", () => {
@@ -186,7 +191,10 @@ describe("Test Tracer", () => {
 					options: {
 						a: 5
 					}
-				}]
+				}],
+				defaultTags: jest.fn(() => ({
+					a: 5
+				}))
 			});
 
 			tracer.init();
@@ -204,6 +212,10 @@ describe("Test Tracer", () => {
 			expect(fakeExporter.init).toBeCalledTimes(2);
 			expect(fakeExporter.init).toHaveBeenNthCalledWith(1, tracer);
 			expect(fakeExporter.init).toHaveBeenNthCalledWith(2, tracer);
+
+			expect(tracer.defaultTags).toEqual({ a: 5 });
+			expect(tracer.opts.defaultTags).toHaveBeenCalledTimes(1);
+			expect(tracer.opts.defaultTags).toHaveBeenCalledWith(tracer);
 		});
 
 	});